Add type definitions
This commit add type definitions for `gulp-typedoc`. These type
definitions are more complete than the ones from DT. They export all
the types (`Options` and `Logger`), are documented and tested.

The tests try to compile a file in `noEmit` mode to verify that it
type-checks. This adds the new devDevdependencies `@types/node` and
`typescript`.

This is a semver minor change (feature).

- See rogierschouten/gulp-typedoc#20